+++ apps/backend/src/donations/dtos/create-donation.domain.ts
@@ -1,13 +1,7 @@
-export const donationTypes = ['one_time', 'recurring'] as const;
-export type DonationType = (typeof donationTypes)[number];
-export const recurringIntervals = [
-  'monthly',
-  'bimonthly',
-  'quarterly',
-  'annually',
-  'weekly',
-] as const;
-export type NormalizedInterval = (typeof recurringIntervals)[number];
+import {
+  DonationType,
+  NormalizedInterval,
+} from '../../util/donations/donations.util';
 
 export class CreateDonationDTO {
   firstName: string;
@@ -27,18 +21,4 @@ export class CreateDonationDTO {
   dedicationMessage?: string | null;
 
   showDedicationPublicly?: boolean = false;
-
-  private static normalizeInterval(
-    input: string | null,
-  ): NormalizedInterval | null {
-    if (!input) {
-      return null;
-    }
-
-    const normalized = input.toLowerCase().trim();
-
-    return (recurringIntervals as readonly string[]).includes(normalized)
-      ? (normalized as NormalizedInterval)
-      : null;
-  }
 }
